Moved logic into save dataset modal
